is_leap_year refactor: Simplify logical returns
The current implementation uses the following pattern (twice):

```python
if condition:
    return True
else:
    return False
```

Given that `condition` itself is a Boolean value (i.e. either `True` or
`False`), this essentially boils down to either

```python
if True:
    return True
```

or

```python
if False:
    return False
```

depending on the actual value of `condition`.

This commit simplifies that return logic by directly returning the value
of `condition` instead:

```python
return condition
````